+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## v1.2.2
 
 ### Added
-  - In `equilibrium_solve!()`, we have now two additional inputs: `verbose` and `abstoly` to control the secant method for finding vacancy energy levels
+  - In `equilibrium_solve!()`, we have now two additional inputs: `verbose` and `yabstol` to control the secant method for finding vacancy energy levels
 
 ## v1.2.1